# perf: `offsetToNeedle` caching improvement