--- a/package.json
+++ b/package.json
@@ -20,6 +20,7 @@
   "scripts": {
     "dev": "tsup --watch --sourcemap",
     "bench": "node --experimental-strip-types --cpu-prof --cpu-prof-dir=./bench-profile ./benchmark/run.ts",
+    "bench:ci": "node --experimental-strip-types ./benchmark/run.ts",
     "build": "tsup",
     "test": "vitest",
     "lint": "eslint . --max-warnings 0",