- `checker.ts` goes from 60s to 1s

```
# Before:
Transforming took 1.83s
AST parsing took 0.15s
Coverage remapping took 59.66s

CoverageSummary {
  data: {
    lines: { total: 23198, covered: 0, skipped: 0, pct: 0 },
    statements: { total: 24054, covered: 0, skipped: 0, pct: 0 },
    functions: { total: 3267, covered: 0, skipped: 0, pct: 0 },
    branches: { total: 27805, covered: 0, skipped: 0, pct: 0 },
    branchesTrue: { total: 0, covered: 0, skipped: 0, pct: 'Unknown' }
  }
}

# After
Transforming took 2.05s
AST parsing took 0.14s
Coverage remapping took 1.00s

CoverageSummary {
  data: {
    lines: { total: 23198, covered: 0, skipped: 0, pct: 0 },
    statements: { total: 24054, covered: 0, skipped: 0, pct: 0 },
    functions: { total: 3267, covered: 0, skipped: 0, pct: 0 },
    branches: { total: 27805, covered: 0, skipped: 0, pct: 0 },
    branchesTrue: { total: 0, covered: 0, skipped: 0, pct: 'Unknown' }
  }
}
```



<img width="400" src="https://github.com/user-attachments/assets/7560fafd-869b-422e-8d7c-65202c2b37bb" />

<img width="400" src="https://github.com/user-attachments/assets/19da04b5-1513-4bf1-9859-984b465a411c" />